refactor: move monaco from plugin
